backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
distro==1.9.0
dj-rest-auth==7.0.0
Django==4.2.16
django-allauth==65.2.0
django-environ==0.11.2
djangorestframework==3.15.2
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
numpy==2.0.2
openai==1.54.4
pgvector==0.3.6
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.2
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
